added PerMainWindowInjector
