+++ src/main/java/amidst/Application.java
@@ -7,6 +7,7 @@ import amidst.documentation.CalledOnlyBy;
 import amidst.documentation.NotThreadSafe;
 import amidst.gui.license.LicenseWindow;
 import amidst.gui.main.MainWindow;
+import amidst.gui.main.MainWindowDialogs;
 import amidst.gui.main.UpdatePrompt;
 import amidst.gui.profileselect.ProfileSelectWindow;
 import amidst.mojangapi.MojangApi;
@@ -14,7 +15,7 @@ import amidst.mojangapi.MojangApi;
 @NotThreadSafe
 public class Application {
 	private final MojangApi mojangApi;
-	private final Factory1<MainWindow, UpdatePrompt> noisyUpdatePromptFactory;
+	private final Factory1<MainWindowDialogs, UpdatePrompt> noisyUpdatePromptFactory;
 	private final Factory0<UpdatePrompt> silentUpdatePromptFactory;
 	private final Factory0<MainWindow> mainWindowFactory;
 	private final Factory0<ProfileSelectWindow> profileSelectWindowFactory;
@@ -26,7 +27,7 @@ public class Application {
 	@CalledOnlyBy(AmidstThread.EDT)
 	public Application(
 			MojangApi mojangApi,
-			Factory1<MainWindow, UpdatePrompt> noisyUpdatePromptFactory,
+			Factory1<MainWindowDialogs, UpdatePrompt> noisyUpdatePromptFactory,
 			Factory0<UpdatePrompt> silentUpdatePromptFactory,
 			Factory0<MainWindow> mainWindowFactory,
 			Factory0<ProfileSelectWindow> profileSelectWindowFactory,
@@ -50,8 +51,8 @@ public class Application {
 	}
 
 	@CalledOnlyBy(AmidstThread.EDT)
-	public void checkForUpdates(MainWindow mainWindow) {
-		noisyUpdatePromptFactory.create(mainWindow).check();
+	public void checkForUpdates(MainWindowDialogs dialogs) {
+		noisyUpdatePromptFactory.create(dialogs).check();
 	}
 
 	@CalledOnlyBy(AmidstThread.EDT)
